requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
beautifulsoup4>=4.12.0
aiohttp>=3.8.0
lxml>=4.9.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
asyncio>=3.4.3
selenium>=4.15.0  
scrapy>=2.11.0    
pytest>=7.4.0
black>=23.7.0
flake8>=6.0.0
matplotlib>=3.7.0  
seaborn>=0.12.0    
plotly>=5.17.0     
openpyxl>=3.1.0
xlsxwriter>=3.1.0
ratelimit>=2.2.1
cachetools>=5.3.0
nltk>=3.8.0        
spacy>=3.7.0       
cryptography>=41.0.0
pyyaml>=6.0.0
google-ads>=22.1.0  
tweepy>=4.14.0      
facebook-sdk>=3.1.0 
asyncpg>=0.28.0     
redis>=5.0.0        
loguru>=0.7.0
sentry-sdk>=1.32.0  